Move controllers into packages
